# openapi.json
{
  "components": {
    "schemas": {
...
      "AssetDetail": {
        "properties": {
          "asset": {
            "$ref": "#/components/schemas/ScreenerItem"
          },
          "price_history": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Price History",
            "type": "array"
          },
          "score_history": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Score History",
            "type": "array"
          },
          "score_trace": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScoreTrace"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "asset",
          "price_history",
          "score_history"
        ],
        "title": "AssetDetail",
        "type": "object"
      },
...
      "BacktestCreate": {
        "properties": {
          "sleeve_weights_bps": {
            "$ref": "#/components/schemas/SleeveWeights"
          }
        },
        "title": "BacktestCreate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Quant Trend Lab API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/assets/{asset_id}": {
      "get": {
        "operationId": "asset_detail_api_v1_assets__asset_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Asset Detail",
        "tags": [
          "research"
        ]
      }
    },
    "/api/v1/backtests": {
      "post": {
        "operationId": "create_backtest_api_v1_backtests_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BacktestCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Backtest",
        "tags": [
          "backtests"
        ]
      }
    },
    "/api/v1/backtests/{run_id}": {
      "get": {
        "operationId": "get_backtest_api_v1_backtests__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Backtest",
        "tags": [
          "backtests"
        ]
      }
    },
    "/api/v1/backtests/{run_id}/artifacts": {
      "get": {
        "operationId": "get_artifacts_api_v1_backtests__run_id__artifacts_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ArtifactResponse"
                  },
                  "title": "Response Get Artifacts Api V1 Backtests  Run Id  Artifacts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Artifacts",
        "tags": [
          "backtests"
        ]
      }
    },
    "/api/v1/forward/accounts": {
      "get": {
        "operationId": "list_forward_accounts_api_v1_forward_accounts_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForwardAccountsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Forward Accounts",
        "tags": [
          "forward-research"
        ]
      },
      "post": {
        "operationId": "create_forward_account_api_v1_forward_accounts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForwardAccountCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForwardAccountResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Forward Account",
        "tags": [
          "forward-research"
        ]
      }
    },
...
    "/api/v1/forward/accounts/{account_id}": {
      "get": {
        "operationId": "get_forward_account_api_v1_forward_accounts__account_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "title": "Account Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForwardAccountResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Forward Account",
        "tags": [
          "forward-research"
        ]
      }
    },
    "/api/v1/forward/accounts/{account_id}/activity": {
      "get": {
        "operationId": "forward_account_activity_api_v1_forward_accounts__account_id__activity_get",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "title": "Account Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForwardActivityResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Forward Account Activity",
        "tags": [
          "forward-research"
        ]
      }
    },
    "/api/v1/forward/accounts/{account_id}/archive": {
      "post": {
        "operationId": "archive_forward_account_api_v1_forward_accounts__account_id__archive_post",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "title": "Account Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForwardAccountResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Archive Forward Account",
        "tags": [
          "forward-research"
        ]
      }
    },
...
  }
}